Move parser to a standlone package
